Switched User to the DB, fixed the save method of Conference, and changed the ids in the fixture to -1 to make sure they won't overwrite other fixtures on save

+++ conf/evolutions/default/1.sql
@@ -38,7 +38,7 @@ CREATE TABLE User (
     email     VARCHAR(255) NOT NULL,
     lab       bigint NOT NULL REFERENCES Lab(id),
     hashedPass VARCHAR(255),
-    role      smallint,
+    role      int,
     rememberMeToken VARCHAR(255),
     PRIMARY KEY (id)
 )
